+++ spec/draught/segment/curve/builder_spec.rb
@@ -1,12 +1,11 @@
-require 'draught/curve_segment_builder'
+require 'draught/segment/curve/builder'
 require 'draught/world'
 
-module Draught
-  RSpec.describe CurveSegmentBuilder do
-    let(:world) { World.new }
+module Draught::Segment
+  RSpec.describe Curve::Builder do
+    let(:world) { Draught::World.new }
     subject { described_class.new(world) }
 
-    let(:world) { World.new }
     let(:start_point) { world.point.zero }
     let(:end_point) { world.point.new(4,0) }
     let(:control_point_1) { world.point.new(1,2) }
@@ -15,9 +14,9 @@ module Draught
       end_point: end_point, control_point_1: control_point_1,
       control_point_2: control_point_2
     } }
-    let(:cubic_bezier) { CubicBezier.new(world, cubic_opts) }
+    let(:cubic_bezier) { Draught::CubicBezier.new(world, cubic_opts) }
     let(:expected_curve_segment) {
-      CurveSegment.new(world, start_point: start_point, cubic_bezier: cubic_bezier)
+      Curve.new(world, start_point: start_point, cubic_bezier: cubic_bezier)
     }
 
     describe "building a Curve Segment from a hash" do
@@ -36,7 +35,7 @@ module Draught
       end
 
       context "handling Metadata" do
-        let(:metadata) { Metadata::Instance.new(name: 'name') }
+        let(:metadata) { Draught::Metadata::Instance.new(name: 'name') }
         let(:path) { world.path.new(points: [start_point, cubic_bezier], metadata: metadata) }
 
         context "when given a start_point and cubic_bezier" do
@@ -44,7 +43,7 @@ module Draught
             subject.build(start_point: start_point, cubic_bezier: cubic_bezier, metadata: metadata)
           }
 
-          specify "produces a CurveSegment with the correct Metadata" do
+          specify "produces a Curve Segment with the correct Metadata" do
             expect(built_curve_segment.metadata).to be(metadata)
           end
         end
@@ -59,7 +58,7 @@ module Draught
             })
           }
 
-          specify "produces a CurveSegment with the correct Metadata" do
+          specify "produces a Curve Segment with the correct Metadata" do
             expect(built_curve_segment.metadata).to be(metadata)
           end
         end
@@ -67,7 +66,7 @@ module Draught
     end
 
     context "building a Curve Segment from a two-item Path" do
-      it "generates the CurveSegment correctly" do
+      it "generates the Curve Segment correctly" do
         path = world.path.new(points: [start_point, cubic_bezier])
 
         expect(subject.from_path(path)).to eq(expected_curve_segment)
@@ -106,11 +105,11 @@ module Draught
       end
 
       context "handling Metadata" do
-        let(:metadata) { Metadata::Instance.new(name: 'name') }
+        let(:metadata) { Draught::Metadata::Instance.new(name: 'name') }
         let(:path) { world.path.new(points: [start_point, cubic_bezier], metadata: metadata) }
         let(:built_curve_segment) { subject.from_path(path) }
 
-        specify "produces a CurveSegment with the correct Metadata" do
+        specify "produces a Curve Segment with the correct Metadata" do
           expect(built_curve_segment.metadata).to be(metadata)
         end
       end
