+++ spec/draught/segment/line/builder_spec.rb
@@ -1,14 +1,14 @@
-require 'draught/line_segment_builder'
+require 'draught/segment/line/builder'
 require 'draught/world'
 
-module Draught
-  RSpec.describe LineSegmentBuilder do
-    let(:world) { World.new }
-    let(:metadata) { Metadata::Instance.new(name: 'name') }
+module Draught::Segment
+  RSpec.describe Line::Builder do
+    let(:world) { Draught::World.new }
+    let(:metadata) { Draught::Metadata::Instance.new(name: 'name') }
 
     subject { described_class.new(world) }
 
-    describe "building a LineSegment between two Points" do
+    describe "building a Line between two Points" do
       let(:finish) { world.point.new(4,4) }
       let(:line_segment) { subject.build(end_point: finish, metadata: metadata) }
 
@@ -32,42 +32,42 @@ module Draught
         expect(line_segment.metadata).to be(metadata)
       end
 
-      specify "a LineSegment at 0º should have radians == 0" do
+      specify "a Line at 0º should have radians == 0" do
         line_segment = subject.build(end_point: world.point.new(10,0))
 
         expect(line_segment.radians).to be_within(0.0001).of(0)
       end
 
       context "angles < 90º" do
-        it "copes with a LineSegment of angle ~89.9º" do
+        it "copes with a Line of angle ~89.9º" do
           line_segment = subject.build(end_point: world.point.new(0.007,4))
 
           expect(line_segment.length).to be_within(0.01).of(4)
           expect(line_segment.radians).to be_within(0.0001).of(deg_to_rad(89.9))
         end
 
-        it "copes with a LineSegment of angle 60º" do
+        it "copes with a Line of angle 60º" do
           line_segment = subject.build(end_point: world.point.new(2.31,4))
 
           expect(line_segment.length).to be_within(0.01).of(4.62)
           expect(line_segment.radians).to be_within(0.001).of(deg_to_rad(60))
         end
 
-        it "copes with a LineSegment of angle 45º" do
+        it "copes with a Line of angle 45º" do
           line_segment = subject.build(end_point: world.point.new(4,4))
 
           expect(line_segment.length).to be_within(0.01).of(5.66)
           expect(line_segment.radians).to be_within(0.0001).of(deg_to_rad(45))
         end
 
-        it "copes with a LineSegment of angle 30º" do
+        it "copes with a Line of angle 30º" do
           line_segment = subject.build(end_point: world.point.new(4,2.31))
 
           expect(line_segment.length).to be_within(0.01).of(4.62)
           expect(line_segment.radians).to be_within(0.001).of(deg_to_rad(30))
         end
 
-        it "copes with a LineSegment of angle ~1º" do
+        it "copes with a Line of angle ~1º" do
           line_segment = subject.build(end_point: world.point.new(4,0.0699))
 
           expect(line_segment.length).to be_within(0.01).of(4)
@@ -76,42 +76,42 @@ module Draught
       end
 
       context "angles >= 90º" do
-        it "copes with a LineSegment of angle 90º" do
+        it "copes with a Line of angle 90º" do
           line_segment = subject.build(end_point: world.point.new(0,4))
 
           expect(line_segment.length).to be_within(0.01).of(4)
           expect(line_segment.radians).to eq(deg_to_rad(90))
         end
 
-        it "copes with a LineSegment of angle < 180º" do
+        it "copes with a Line of angle < 180º" do
           line_segment = subject.build(end_point: world.point.new(-4,4))
 
           expect(line_segment.length).to be_within(0.01).of(5.66)
           expect(line_segment.radians).to be_within(0.0001).of(deg_to_rad(135))
         end
 
-        it "copes with a LineSegment of angle 180º" do
+        it "copes with a Line of angle 180º" do
           line_segment = subject.build(end_point: world.point.new(-4,0))
 
           expect(line_segment.length).to be_within(0.01).of(4)
           expect(line_segment.radians).to eq(deg_to_rad(180))
         end
 
-        it "copes with a LineSegment of angle < 270º" do
+        it "copes with a Line of angle < 270º" do
           line_segment = subject.build(end_point: world.point.new(-4,-4))
 
           expect(line_segment.length).to be_within(0.01).of(5.66)
           expect(line_segment.radians).to be_within(0.0001).of(deg_to_rad(225))
         end
 
-        it "copes with a LineSegment of angle 270º" do
+        it "copes with a Line of angle 270º" do
           line_segment = subject.build(end_point: world.point.new(0,-4))
 
           expect(line_segment.length).to be_within(0.01).of(4)
           expect(line_segment.radians).to eq(deg_to_rad(270))
         end
 
-        it "copes with a LineSegment of angle < 360º" do
+        it "copes with a Line of angle < 360º" do
           line_segment = subject.build(end_point: world.point.new(4,-4))
 
           expect(line_segment.length).to be_within(0.01).of(5.66)
@@ -120,7 +120,7 @@ module Draught
       end
     end
 
-    describe "generating horizontal LineSegment objects" do
+    describe "generating horizontal Line objects" do
       specify "a line_segment of width N is like a Path with points at (0,0) and (N,0)" do
         expected = world.path.build { |p| p << world.point.zero << world.point.new(10, 0) }
 
@@ -134,7 +134,7 @@ module Draught
       end
     end
 
-    describe "generating vertical LineSegment objects" do
+    describe "generating vertical Line objects" do
       specify "a line_segment of height N is like a Path with points at (0,0) and (0,N)" do
         expected = world.path.build { |p| p << world.point.zero << world.point.new(0, 10) }
 
@@ -148,98 +148,29 @@ module Draught
       end
     end
 
-    describe "generating LineSegment objects of a given length and angle" do
-      def fixture_path
-        (Pathname.new(__dir__)/'../fixtures/line-segment-extension.svg')
-      end
-
-      def parse_fixture
-        parser = Parser::SVG.new(world, fixture_path.open('r:utf-8'))
-        parser.parse!
-      end
-
-      def line(angle)
-        svg = parse_fixture
-        world.line_segment.from_path(svg.paths.find { |path| path.name == "a-#{angle}" })
-      end
+    describe "generating Line objects of a given length and angle", :svg_fixture do
+      let(:tolerance) { Draught::Tolerance.new(0.001) } # Affinity Designer only goes to 3 d.p.
+      let(:world) { Draught::World.new(tolerance) }
 
       def deg_to_rad(degrees)
+        degrees = degrees.is_a?(String) ? Integer(degrees, 10) : degrees
         degrees * (Math::PI / 180)
       end
 
-      let(:length) { 5.656854 }
-
-      it "copes with a line_segment of angle 0º" do
-        expected = subject.build(end_point: world.point.new(length,0))
-
-        expect(subject.build(length: length, radians: 0)).to eq(expected)
-      end
-
-      it "copes with a LineSegment of angle 45º" do
-        expected = subject.build(end_point: world.point.new(4,4))
-
-        expect(subject.build({
-          length: length, radians: deg_to_rad(45)
-        })).to approximate(expected).within(0.00001)
-      end
-
-      it "copes with a LineSegment of angle 90º" do
-        expected = subject.build(end_point: world.point.new(0,length))
-
-        expect(subject.build(length: length, radians: deg_to_rad(90))).to eq(expected)
-      end
-
-      it "copes with a LineSegment of angle < 180º" do
-        expected = subject.build(end_point: world.point.new(-4,4))
-
-        expect(subject.build({
-          length: length, radians: deg_to_rad(135)
-        })).to approximate(expected).within(0.00001)
-      end
-
-      it "copes with a LineSegment of angle 180º" do
-        expected = subject.build(end_point: world.point.new(-length,0))
-
-        expect(subject.build(length: length, radians: deg_to_rad(180))).to eq(expected)
-      end
-
-      it "copes with a LineSegment of angle < 270º" do
-        expected = line(210)
-
-        expected = subject.build(end_point: expected.end_point)
-
-        expect(subject.build({
-          length: 100, radians: deg_to_rad(210)
-        })).to approximate(expected).within(0.001)
-      end
-
-      it "copes with a LineSegment of angle < 270º" do
-        expected = subject.build(end_point: world.point.new(-4,-4))
-
-        expect(subject.build({
-          length: length, radians: deg_to_rad(225)
-        })).to approximate(expected).within(0.00001)
-      end
-
-      it "copes with a LineSegment of angle 270º" do
-        expected = subject.build(end_point: world.point.new(0,-length))
-
-        expect(subject.build({
-          length: length, radians: deg_to_rad(270)
-        })).to eq(expected)
-      end
-
-      it "copes with a LineSegment of angle < 360º" do
-        expected = subject.build(end_point: world.point.new(4,-4))
-
-        expect(subject.build({
-          length: length, radians: deg_to_rad(315)
-        })).to approximate(expected).within(0.00001)
+      svg_fixture('line-segment-extension.svg') {
+        fetch_grouped(expected: /a-([0-9]+)$/)
+        map_paths { |world, path| world.line_segment.from_path(path) }
+      }.each do |world, angle, expected|
+        specify "correctly builds a line of angle #{angle}º" do
+          expect(subject.build(length: 100, radians: deg_to_rad(angle))).to eq(expected)
+        end
       end
 
       context "ludicrous angles" do
+        let(:length) { 5.656854 }
+
         it "treats a 360º angle as 0º" do
-          expected = subject.build(end_point: world.point.new(length,0))
+          expected = subject.build(end_point: world.point.new(length, 0))
 
           expect(subject.build(length: length, radians: deg_to_rad(360))).to eq(expected)
         end
@@ -276,7 +207,7 @@ module Draught
       context "handling Metadata" do
         let(:line_segment) {
           subject.build({
-            length: length, radians: deg_to_rad(45),
+            length: 50, radians: deg_to_rad(45),
             metadata: metadata
           })
         }
@@ -287,23 +218,23 @@ module Draught
       end
     end
 
-    describe "generating LineSegment objects that don't start at 0,0" do
-      it "can generate a LineSegment from points" do
+    describe "generating Line objects that don't start at 0,0" do
+      it "can generate a Line from points" do
         line_segment = subject.build(start_point: world.point.new(1,1), end_point: world.point.new(5,5))
 
         expect(line_segment.radians).to be_within(0.00001).of(Math::PI/4)
         expect(line_segment.length).to be_within(0.01).of(5.66)
       end
 
-      it "can generate a LineSegment from angle/length and start point" do
+      it "can generate a Line from angle/length and start point" do
         line_segment = subject.build(start_point: world.point.new(1,1), radians: Math::PI/4, length: 5.656854)
 
         expect(line_segment).to approximate(world.path.new(points: [world.point.new(1,1), world.point.new(5,5)])).within(0.00001)
       end
     end
 
-    describe "building a LineSegment from a two-item Path" do
-      it "generates the LineSegment correctly" do
+    describe "building a Line from a two-item Path" do
+      it "generates the Line correctly" do
         path = world.path.new(points: [world.point.zero, world.point.new(4,4)])
 
         expect(subject.from_path(path)).to eq(path)
@@ -333,11 +264,11 @@ module Draught
       end
     end
 
-    describe "building a LineSegment from two points" do
+    describe "building a Line from two points" do
       let(:p1) { world.point.zero }
       let(:p2) { world.point.new(4,4) }
 
-      it "generates the LineSegment correctly" do
+      it "generates the Line correctly" do
         path = world.path.new(points: [p1, p2])
 
         expect(subject.from_to(p1, p2)).to eq(path)
